Implemented scene node management and scene creation.

+++ modules/Scene/src/Nodes/SceneNode.c
@@ -22,6 +22,7 @@
 #include <DeepSea/Core/Assert.h>
 #include <DeepSea/Core/Atomic.h>
 #include <DeepSea/Core/Error.h>
+#include <DeepSea/Core/Log.h>
 #include <string.h>
 
 size_t dsSceneNode_drawListsAllocSize(const char** drawLists, uint32_t drawListCount)
@@ -39,9 +40,16 @@ bool dsSceneNode_initialize(dsSceneNode* node, dsAllocator* allocator,
 	dsSceneNodeType type, const char** drawLists, uint32_t drawListCount,
 	dsDestroySceneNodeFunction destroyFunc)
 {
-	if (!node || !dsAllocator_keepPointer(allocator) || !destroyFunc)
+	if (!node || !allocator || !destroyFunc)
 	{
 		errno = EINVAL;
+		return false;
+	}
+
+	if (!allocator->freeFunc)
+	{
+		errno = EINVAL;
+		DS_LOG_ERROR(DS_SCENE_LOG_TAG, "Scene node allocator must support freeing memory.");
 		return false;
 	}
 
@@ -56,11 +64,62 @@ bool dsSceneNode_initialize(dsSceneNode* node, dsAllocator* allocator,
 	node->treeNodeCount = 0;
 	node->maxTreeNodes = 0;
 	node->refCount = 1;
+	node->nextChildID = 0;
 	node->destroyFunc = destroyFunc;
 	return true;
 }
 
-bool dsSceneNode_addChild(dsSceneNode* node, dsSceneNode* child)
+uint32_t dsSceneNode_addChild(dsSceneNode* node, dsSceneNode* child)
+{
+	if (!node || !child)
+	{
+		errno = EINVAL;
+		return DS_NO_SCENE_NODE;
+	}
+
+	uint32_t index = node->childCount;
+	if (!DS_RESIZEABLE_ARRAY_ADD(node->allocator, node->children, node->childCount,
+			node->maxChildren, 1))
+	{
+		return DS_NO_SCENE_NODE;
+	}
+
+	dsSceneNodeChildRef* childRef = node->children + index;
+	childRef->node = dsSceneNode_addRef(child);
+	childRef->childID = node->nextChildID++;
+	if (!dsSceneTreeNode_buildSubtree(node, childRef))
+	{
+		dsSceneNode_freeRef(child);
+		--node->childCount;
+		return DS_NO_SCENE_NODE;
+	}
+
+	return childRef->childID;
+}
+
+bool dsSceneNode_removeChildIndex(dsSceneNode* node, uint32_t childIndex)
+{
+	if (!node)
+	{
+		errno = EINVAL;
+		return false;
+	}
+
+	if (childIndex >= node->childCount)
+	{
+		errno = EINDEX;
+		return false;
+	}
+
+	dsSceneNodeChildRef* child = node->children + childIndex;
+	dsSceneTreeNode_removeSubtree(node, child->node, child->childID);
+	// Order shouldn't matter, so put the last item in this spot for constant-time removal.
+	node->children[childIndex] = node->children[node->childCount - 1];
+	--node->childCount;
+	return true;
+}
+
+bool dsSceneNode_removeChildNode(dsSceneNode* node, dsSceneNode* child, uint32_t childID)
 {
 	if (!node || !child)
 	{
@@ -68,22 +127,35 @@ bool dsSceneNode_addChild(dsSceneNode* node, dsSceneNode* child)
 		return false;
 	}
 
-	uint32_t index = node->childCount;
-	if (!DS_RESIZEABLE_ARRAY_ADD(node->allocator, node->children, node->childCount,
-			node->maxChildren, 1))
+	dsSceneTreeNode_removeSubtree(node, child, childID);
+	for (uint32_t i = node->childCount; i-- > 0;)
 	{
-		return false;
-	}
+		dsSceneNodeChildRef* childRef = node->children + i;
+		if (childRef->node != child ||
+			(childID != DS_NO_SCENE_NODE && childRef->childID != childID))
+		{
+			continue;
+		}
 
-	node->children[index] = dsSceneNode_addRef(child);
-	if (!dsSceneTreeNode_buildSubtree(node, child))
-	{
 		dsSceneNode_freeRef(child);
+		node->children[i] = node->children[node->childCount - 1];
 		--node->childCount;
-		return false;
 	}
-
 	return true;
+}
+
+void dsSceneNode_clear(dsSceneNode* node)
+{
+	if (!node)
+		return;
+
+	for (uint32_t i = 0; i < node->childCount; ++i)
+	{
+		dsSceneNode* child = node->children[i].node;
+		dsSceneTreeNode_removeSubtree(node, child, DS_NO_SCENE_NODE);
+		dsSceneNode_freeRef(child);
+	}
+	node->childCount = 0;
 }
 
 dsSceneNode* dsSceneNode_addRef(dsSceneNode* node)
@@ -103,13 +175,10 @@ void dsSceneNode_freeRef(dsSceneNode* node)
 	if (DS_ATOMIC_FETCH_ADD32(&node->refCount, -1) != 1)
 		return;
 
-	for (uint32_t i = 0; i < node->childCount; ++i)
-	{
-		dsSceneNode* child = node->children[i];
-		dsSceneTreeNode_removeSubtree(node, child);
-		dsSceneNode_freeRef(child);
-	}
+	dsSceneNode_clear(node);
+	DS_VERIFY(dsAllocator_free(node->allocator, node->children));
+	DS_VERIFY(dsAllocator_free(node->allocator, node->treeNodes));
 
-	DS_ASSERT(node->destroyFunc);
-	node->destroyFunc(node);
+	if (node->destroyFunc)
+		node->destroyFunc(node);
 }
